Cleanup for a new start
